+++ src/Model/DataProvider/DataProviderIssueData.php
@@ -21,6 +21,8 @@ class DataProviderIssueData
         public ?\DateTimeInterface $fetchTime,
         public ?string $url,
         public ?\DateTimeInterface $sourceModifiedDate,
+        public ?string $versionId,
+        public bool $disableModifiedAtCheck = false,
     ) {
     }
 }